app/*: implement zap logging and errors (#81)
This follows on the improve errors and logging proposal. It adds `app/log` and `app/errors` packages that should be used for all errors and logging in the application.

It introduces contextual logging via `log.WithCtx`.

It introduces structured errors (with stack traces) via `errors.New` and `errors.Wrap`.

It also introduces the `z` package which wraps `zap.Field` for a more concise API and to support multi-zap-fields required for structured errors.

It also introduces an opinionated first version of a console logger.