cluster/leadercast: fix synchronisation bug (#123)
Fiixes synchronisation bug in leadercast with sync.Cond. If multiple goroutines were waiting on the condition, only a single one is woken up by sync.Cond.Signal(), if the message that arrived was for the other goroutine, then one of the two goroutines will get stuck forever. All goroutines should be woken up to check if it was their message that arrived, not just one.

Fix by using mutex and a shared channel map instead.