core/validatorapi: decrease logging for 400 errors (#402)
As per golang guidelines, errors are for actionable problems. 400 responses are expected so shouldn't be errors. For example, when a validator isn't active yet, lookup validators returns 400. Or if a validator doesn't have a duty for a slot, 400 are expected. Decrease logging level to debug. Since if it really is an issue, the VC logs will contain the required error logs.

We do have metrics for all responses, so they are tracked.

category: refactor
ticket: none