p2p: decrease circuit relay reservation frequency (#892)
This reduces libp2p circuit relay reservations to the required minimum. Previously, we re-reserved on "client connection timeout" of 2s which is not required since server-side reservations remain open when client-side connections are recycled. This resulted in multiple open reservations which quickly results in "reservation-rejected" errors. 

category: bug
ticket: none