testutil/validatormock: add pseudo sync committee contribution flow (#1172)
Adds a pseudo sync committee contribution flow. It is pseudo code in a test file, it compiles, but is never called. It aims to show how we are going to implement the sync committee flow from the VC perspective. It also proposes a new endpoint to be introduces to add support for middle DVT where is `is_sync_committee_aggregator` cannot be calculated locally since `get_sync_committee_selection_proof` produces a partial signature.

category: misc
ticket: none