p2p: do periodic pings (#1418)
Refactor pings to be periodic (once per second). Currently we ping as fast as we can. In cloud clusters with low latency, this can go up to thousands of pings per second.

Also refactor the connection duration histogram into a connection counter as that is much simpler and easier to use in dashboards.

category: refactor
ticket: none