app/log: fix missing stacktraces (#1718)
Fixes missing error/warn stack traces when logged from docker built binaries. This is due to `formatZapStack` that assumes the source code was located in a `.../charon/` folder when built. Not sure how to fix that properly, so the workaround is to rename the folder used when building the source.  

category: bug
ticket: none