integration: wait for beaconmock to be up before starting teku
It might happen that the beaconmock goroutine gets scheduled much after the teku startup one, leading to Teku exiting with error, and the test failing.

Ensure this doesn't happen by checking that beaconmock is up before running the teku container: if it doesn't become available in 10s (that is, HTTP requests to it doesn't error), error the test.