testutil/verifypr: limit PR titles to 60 chars (#2056)
Limit PR titles to 60 chars. This drives the use of succinct PR titles and more verbose bodies. Commit and PR titles should be glanceable. It should be easy to glance over a long list of PRs and quickly identify and find things "at a glance". This requires short titles. Details about what changed, the types involved, the constants changed all need to be in the body. 

category: misc
ticket: none