app/eth2wrap: implement and wire validator cache (#2121)
This is a second attempt at caching active validators. Instead of caching the `eth2client.ValidatorProvider` endpoints which return full `Validator` objects that are not actually immutable by epoch, rather implement a new set of client endpoint specifically for cached immutable active validators by current epoch that only return the validator index and pubkey which are immutable. The only race condition is the fact that we query this by "head", so could be for another slot that what we think is current, although this was already the case previously, so no new races are introdcued.

This should significantly decrease BN load, especially if using the validatormock.  

category: feature
ticket: #1396