eth2util/keystore: order keystore names by their index
Fixes an error in cmd/combine where the i'th combined key was different than excepted if the amount of validator is greater than 10.

filepath.Glob() orders its input lexicographically, so that the following output is obtained:

├── keystore-insecure-0.json
├── keystore-insecure-0.txt
├── keystore-insecure-1.json
├── keystore-insecure-1.txt
├── keystore-insecure-10.json
├── keystore-insecure-10.txt
├── keystore-insecure-2.json
├── keystore-insecure-2.txt

In this instance, keystore 10 comes before keystore 2.

cmd/combine uses the lockfile keys order to orchestrate its logic and given the mismatch between that and the LoadKeys() output, a key mismatch error will result.

Ordering the keys path before reading fixes this issue.