p2p: use direct connection for ping (#2230)
The Ping service was reporting wrong latencies. This is because the ping service creates a long lived stream to the peer using the relay connection. It uses relay connection to report metrics even if direct connection is now available. This is resolved by using new stream/connection for each ping.

category: bug
ticket: none