core/consensus: propose returns instance errors (#2249)
Consensus has two "start" methods, `Participate` and `Propose` either of which can run the instance (depending on timer type). This causes problems with tracker since it doens't know which method to track and the fact that `Participate` is called independent of fetcher, means it breaks tracker step ordering logic.

Refactor consensus to always return the instance error from `Propose`, regardless of whether it was started from `Participate` or `Propose`. `Participate` doesn't need to be instrumented, and the order of events in tracker remain consistent.

category: bug
ticket: none