core/validatorapi: swallow non-dv registrations (#2330)
Swallow incorrect non-dv validator registrations. This is a known limitation. Most VCs do not support valid dv-middleware validator registrations (only teku does at this point). But with pre-generated registrations, we do enable builder-api on VCs to enable mev-boost, so that results in these invalid registration being submitted which results in a lot of `unknown public key` errors in both charon as well as errors the VC logs. These errors are just noise, resulting in operators worrying unnecessarily. 

The risk is that we do not identify actual bugs, but that is why we do the debug log (with log filtering).

category: misc
ticket: none