cmd/combine: always load the same set of files
manifest.Load() gives precedence to manifest files instead of locks, so that even in case a charon directory contains both, manifest is always chosen.

Make sure that combine checks the second return value of that method and stop the process if a lock file is loaded instead of a manifest.

TLDR: combined directories must all load manifest files OR lock files.