eth2util/keymanager: dynamically set request timeout (#2586)
Using the `--keymanager-addresses` flag to automatically import keystores fails when creating a cluster with more than a handful of keystores due to the request timing out while the remote signer is importing the keystores.

This PR adjusts the behavior by making the timeout value dependent on the amount of keystores to be imported, setting a timeout of 2 seconds per keystore. This value was chosen rather arbitrarily as a "safe", high enough value based on my experience with importing keystores, this taking around 0.25 - 0.5 seconds per keystore.

Feel free to lower the value as you deem fit. However, the request timing out can be rather annoying to resolve, since the entire `create cluster` process needs to be started over and keystores need to manually be removed from keymanagers where they were already imported. This is why I propose to use a safe, albeit little high, timeout value.

category: bug
ticket: none