app/eth2wrap: fix error comparison for synthetic blocks (#2705)
This PR cherry-picks #2702 onto `main-v0.18`

---

There was a small issue when comparing eth2api errors when searching for signed beacon block. The error returned should be part of eth2wrap multi implementation which was not able to decompose eth2 errors properly. This PR fixes decomposition of eth2 errors and adds a function to check if the relevant field is present as part of the error chain.

category: bug
ticket: #2695