dkg/sync: increase by max 2 steps
Allow step increase by a maximum amount of 2 steps.

Since the sync protocol runs in the context of a distributed system, peer messages might be received in a different order than expected.

This is especially true when the nodes are e.g. writing data to disk.

This change still keeps the monotonicity property of the function: step count will never decrease.