core/aggsigdb: refactor MemDB
Introduce MemDBV2: instead of relying on a complex system based on channels and read/write queries, simply store objects in a `sync/map`.

Simpler to read and potentially faster in systems in which there's a high rate of goroutine contention.

Add the `aggsigdb_v2` feature flag to enable/disable this feature.