cmd/combine: fix multiple runs
Instead of checking if `validator_keys` directory exists at the end of the load manifest loop, do it earlier and exclude potential non-Charon node directories.

Also make so that `--force` deletes the existing keys, if any.

Solves the use case in which the user called `combine` command twice with an input directory that was also used as an output directory.