dkg: increase test timeout for `TestSigsExchange`
`TestSigsExchange` is a network-intensive test, that exchanges `n^n` messages at a given time.

While it runs fine on development machines, even our more powerful self-hosted runners need more time for it to finish properly.

Increase the timeout from 15 seconds to 45 seconds as well as reduce the amount of nodes from 10 to 7, in hope to reduce the chances of false positives failures in our CI.