*: prep for 1.3.2 (#3698)
* app: remove check for unexisting v2 endpoint (#3695)

Before pectra we've checked if v2 endpoint exist or not. We expect post-pectra this endpoint to always exist and we default only to non-existing block, as before.

category: misc
ticket: none

* app: ssz marshal validx (#3697)

Marshal validator index in SSZ versioned attestations, which was omitted before.

This fixes an issue where the nodes that have first received a versioned attestation from a peer and not from their validator client submit an attestation with empty validator index to the beacon node (as the first saved partial signed data is without the validx).

category: bug
ticket: none

* tracker: check inclusion by validx, instead of agg bits (#3696)

Check attestation inclusion in an easier way, by validator index, instead of aggregation bits, which was throwing false positives before.

category: bug
ticket: none