p2p: error on closing canceled stream (#3916)
Functions `SendReceive` and `Send` create a libp2p stream to send a message to a peer that is closed at the end. Upon enabling QUIC an error `close called for canceled stream` start appearing more often which would cause the function to return the error signaling a failure in sending the message. This would trigger a retry mechanism and add lots of logs.

Upon testing, we verified that this retry is not necessary since the destination peer would receive the message anyways even if we failed to close the stream. This should also not cause any leaks since the stream was canceled.

We still do not know who/what is causing the stream to be canceled.
The PR simply aims to reduce unnecessary retries and logs.

category: bug 
ticket: none