core/fetcher: update blinded gauge metric values (#4170)
Previously `core_fetcher_proposal_blinded` had values 0 (local) or 1 (blinded). This is problematic because the default value for the gauge is also zero so when a node starts/restarts it reports a 0 value before fetching any block which makes monitoring more complicated.

This updates the values to be 2 (local) and 1 (blinded)

category: bug
ticket: none